api/requirements.txt
spacy==3.4.1
unidecode==1.3.6
quantulum3==0.7.10
flask==2.2.2
flask-cors==3.0.10
mysql_connector_python==8.0.30
